Rename PaymentRequestChecker to provider
