+++ src/Sylius/Bundle/CoreBundle/PaymentRequest/CommandHandler/Offline/CapturePaymentRequestHandler.php
@@ -13,23 +13,23 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Offline;
 
-use Sylius\Bundle\CoreBundle\PaymentRequest\Checker\PaymentRequestIntegrityCheckerInterface;
 use Sylius\Bundle\CoreBundle\PaymentRequest\Command\Offline\CapturePaymentRequest;
 use Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Offline\CaptureProcessorInterface;
+use Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface;
 use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 /** @experimental */
 final class CapturePaymentRequestHandler implements MessageHandlerInterface
 {
     public function __construct(
-        private readonly PaymentRequestIntegrityCheckerInterface $paymentRequestIntegrityChecker,
-        private readonly CaptureProcessorInterface $offlineCaptureProcessor
+        private readonly PaymentRequestProviderInterface $paymentRequestProvider,
+        private readonly CaptureProcessorInterface $offlineCaptureProcessor,
     ) {
     }
 
     public function __invoke(CapturePaymentRequest $capturePaymentRequest): void
     {
-        $paymentRequest = $this->paymentRequestIntegrityChecker->check($capturePaymentRequest);
+        $paymentRequest = $this->paymentRequestProvider->provide($capturePaymentRequest);
 
         $this->offlineCaptureProcessor->process($paymentRequest);
     }
