+++ src/Sylius/Bundle/CoreBundle/PaymentRequest/Provider/PaymentRequestProvider.php
@@ -11,37 +11,28 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Checker;
+namespace Sylius\Bundle\CoreBundle\PaymentRequest\Provider;
 
 use Sylius\Bundle\CoreBundle\PaymentRequest\Command\PaymentRequestHashAwareInterface;
-use Sylius\Component\Core\Model\PaymentMethodInterface;
-use Sylius\Component\Payment\Exception\NullGatewayConfigException;
 use Sylius\Component\Payment\Exception\PaymentRequestNotFoundException;
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 use Sylius\Component\Payment\Repository\PaymentRequestRepositoryInterface;
 
-final class PaymentRequestIntegrityChecker implements PaymentRequestIntegrityCheckerInterface
+final class PaymentRequestProvider implements PaymentRequestProviderInterface
 {
+    /**
+     * @param PaymentRequestRepositoryInterface<PaymentRequestInterface> $paymentRequestRepository
+     */
     public function __construct(
         private PaymentRequestRepositoryInterface $paymentRequestRepository,
     ) {
     }
 
-    public function check(PaymentRequestHashAwareInterface $command): PaymentRequestInterface
+    public function provide(PaymentRequestHashAwareInterface $command): PaymentRequestInterface
     {
-        $hash = $command->getHash();
-
-        $paymentRequest = $this->paymentRequestRepository->findOneByHash($hash);
+        $paymentRequest = $this->paymentRequestRepository->find($command->getHash());
         if (null === $paymentRequest) {
             throw new PaymentRequestNotFoundException();
-        }
-
-        /** @var PaymentMethodInterface $paymentMethod */
-        $paymentMethod = $paymentRequest->getMethod();
-
-        $gatewayConfig = $paymentMethod->getGatewayConfig();
-        if (null === $gatewayConfig) {
-            throw new NullGatewayConfigException();
         }
 
         return $paymentRequest;
