+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/checker.xml
@@ -21,10 +21,5 @@
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Checker\PaymentRequestDuplicationCheckerInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Checker\PaymentRequestDuplicationChecker">
             <argument type="service" id="sylius.repository.payment_request" />
         </service>
-
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Checker\PaymentRequestIntegrityCheckerInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Checker\PaymentRequestIntegrityChecker">
-            <argument type="service" id="sylius.repository.payment_request" />
-        </service>
-
     </services>
 </container>
